+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres/tries to adhere to [Semantic Versioning].
 - Add multi-stream chat embed support (big thanks to [@mattroseman](https://github.com/mattroseman) <3)
 - Excluded whispers from phrase detection (big PepoTurkey to Revel)
 - Add channel+title formatting for YouTube embeds (big thanks to [@cantclosevim](https://github.com/cantclosevim) <3)
+- Add Rumble embed formatting
 
 ## [1.7.4] - 2023-02-17